Display parsed ESE motifs in success-alert
This mimicks the way the parsed sequence is displayed in the sequence-
success-alert. Displaying the parsed motifs seems neccessary since the
user would otherwise not know which motifs the dataset contains. (It
has turned out that displaying the file's content in the textarea (=
before upload) is too much trouble, and displaying the parsed data (=
after upload) seems a good alternative.