Set option 'select-by' to default if not specified
Changed order of events: Rather than init the select-by option with its
default it is now set to default only if no value has been specified.
This allows to inform the user that a default value will be used (if
and only if no other value is specified). The parameter description has
been slightly updated to reflect this.